docs: add a status for know issues
As time passes, a known issues can be no longer applicable (because
the component has been fixed, or a workaround ahs been added in the
crosstool-NG scripts/config...).

Some may also be unconfirmed, with no way to easily trigger the error,
or applies to old versions of components...

Add a status for every known issues.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@anciens.enib.fr>